requirements.txt
aiobotocore==2.4.2
aiofiles==22.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
botocore==1.27.59
Brotli==1.0.9
brotlipy==0.7.0
cachelib==0.9.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
comm==0.1.2
cryptography==38.0.4
dash==2.6.0
dash-bootstrap-components==1.2.0
dash-core-components==2.0.0
dash-extensions==0.1.5
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
EditorConfig==0.12.3
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.1.0
Flask-Caching==2.0.0
Flask-Compress==1.11
Flask-Login==0.6.1
flit_core==3.6.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==2.0.0
install==1.3.5
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.1
jmespath==1.0.1
joblib==1.2.0
jsbeautifier==1.14.4
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==8.13.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
platformdirs==3.0.0
plotly==5.6.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==22.0.0
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.27.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3fs==2023.1.0
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
setuptools==65.6.3
sip==6.6.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.3
stack-data==0.6.2
tenacity==8.0.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.9
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.0
wheel==0.37.1
widgetsnbextension==4.0.5
wrapt==1.14.1
y-py==0.5.5
yapf==0.32.0
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.7.0
dash_daq